fixed errors resulting from restructuring the directory structure
