+++ package.json
@@ -5,11 +5,11 @@
   "main": "./dist/punchcards.js",
   "scripts": {
     "postinstall": "bower install && typings install",
-    "tslint": "tslint --config tslint.json ./src/ts/**/*.ts",
+    "tslint": "tslint --config tslint.json ./src/**/*.ts",
     "pretsc": "npm run tslint",
     "tsc": "tsc --project .",
     "predist": "trash dist/ && mkdirp dist",
-    "dist": "browserify --debug build/ts/punchcards.js --outfile dist/punchcards.js --standalone punchcards",
+    "dist": "browserify --debug build/punchcards.js --outfile dist/punchcards.js --standalone punchcards",
     "test": "karma start karma.conf.js",
     "precover": "trash sites/coverage/ && npm run tsc && npm run dist",
     "cover": "npm run test && npm run remap-istanbul-json && npm run remap-istanbul-html",
@@ -17,22 +17,22 @@
     "remap-istanbul-html": "remap-istanbul --input sites/coverage/coverage-remapped.json --output sites/coverage/coverage-remapped --type html --exclude 'node_modules'",
     "pretsdoc": "npm run clean-tsdoc && mkdirp sites/tsdoc",
     "tsdoc": "typedoc --out sites/tsdoc/ .",
-    "clean": "npm run clean-build & npm run clean-dist & npm run clean-cover & npm run clean-sites",
+    "clean": "npm run clean-build & npm run clean-dist & npm run clean-sites",
     "clean-build": "trash build/",
     "clean-dist": "trash dist/",
+    "clean-sites": "npm run clean-cover & npm run clean-demo & npm run clean-tsdoc",
     "clean-cover": "trash sites/coverage/",
-    "clean-sites": "npm run clean-demo & npm run clean-tsdoc",
-    "clean-demo": "trash sites/demo/css/ sites/demo/js/ sites/demo/fonts/ ",
+    "clean-demo": "trash sites/demo/*.js sites/demo/*.css sites/demo/glyphicons-halflings-regular.*",
     "clean-tsdoc": "trash sites/tsdoc/",
     "clean-downloaded": "trash bower_components/ typings/ node_modules/",
     "purge": "npm run clean & npm run clean-downloaded",
     "copy-sites-demo-deps": "npm run copy-sites-demo-js && npm run copy-sites-demo-css && npm run copy-sites-demo-fonts",
-    "copy-sites-demo-css": "cp src/css/punchcards/*.css bower_components/bootstrap/dist/css/bootstrap.min.css bower_components/bootstrap/dist/css/bootstrap-theme.min.css sites/demo/css",
-    "copy-sites-demo-fonts": "cp bower_components/bootstrap/dist/fonts/* sites/demo/fonts/",
-    "copy-sites-demo-js": "cp bower_components/crossfilter/crossfilter.min.js bower_components/moment/min/moment.min.js bower_components/d3/d3.min.js sites/demo/js",
-    "predemo": "npm run clean-demo && mkdirp sites/demo/js sites/demo/css sites/demo/fonts/ && npm run copy-sites-demo-deps",
+    "copy-sites-demo-css": "cp src/*.css bower_components/bootstrap/dist/css/bootstrap.min.css bower_components/bootstrap/dist/css/bootstrap-theme.min.css sites/demo/",
+    "copy-sites-demo-fonts": "cp bower_components/bootstrap/dist/fonts/* sites/demo/",
+    "copy-sites-demo-js": "cp bower_components/crossfilter/crossfilter.min.js bower_components/moment/min/moment.min.js bower_components/d3/d3.min.js sites/demo/",
+    "predemo": "npm run clean-demo && mkdirp sites/demo && npm run copy-sites-demo-deps",
     "demo": "npm run tsc && npm run dist",
-    "postdemo": "cp dist/punchcards.js sites/demo/js/",
+    "postdemo": "cp dist/punchcards.js sites/demo/",
     "everything": "npm run demo && npm run tsdoc && npm run cover"
   },
   "keywords": [
