+++ sites/demo/index.html
@@ -5,21 +5,21 @@
     <meta http-equiv="content-type" content="text/html; charset=utf-8">
     <title>punchcards</title>
 
-    <link rel="stylesheet" type="text/css" href="css/base.css">
-    <link rel="stylesheet" type="text/css" href="css/date-circle.css">
-    <link rel="stylesheet" type="text/css" href="css/date-rect.css">
-    <link rel="stylesheet" type="text/css" href="css/legend.css">
-    <link rel="stylesheet" type="text/css" href="css/weekday-circle.css">
-    <link rel="stylesheet" type="text/css" href="css/weekday-rect.css">
+    <link rel="stylesheet" type="text/css" href="base.css">
+    <link rel="stylesheet" type="text/css" href="date-circle.css">
+    <link rel="stylesheet" type="text/css" href="date-rect.css">
+    <link rel="stylesheet" type="text/css" href="legend.css">
+    <link rel="stylesheet" type="text/css" href="weekday-circle.css">
+    <link rel="stylesheet" type="text/css" href="weekday-rect.css">
 
-    <link rel="stylesheet" type="text/css" href="css/bootstrap.min.css">
+    <link rel="stylesheet" type="text/css" href="bootstrap.min.css">
     <!-- be aware that the theme also needs some fonts -->
-    <link rel="stylesheet" type="text/css" href="css/bootstrap-theme.min.css">
+    <link rel="stylesheet" type="text/css" href="bootstrap-theme.min.css">
 
-    <script type="text/javascript" src="js/punchcards.js"></script>
-    <script type="text/javascript" src="js/crossfilter.min.js"></script>
-    <script type="text/javascript" src="js/moment.min.js"></script>
-    <script type="text/javascript" src="js/d3.min.js"></script>
+    <script type="text/javascript" src="punchcards.js"></script>
+    <script type="text/javascript" src="crossfilter.min.js"></script>
+    <script type="text/javascript" src="moment.min.js"></script>
+    <script type="text/javascript" src="d3.min.js"></script>
 
 
   </head>
@@ -31,8 +31,8 @@
         <div id="punchcard-date-circle" class="punchcard-common punchcard-date-circle"></div>
         <div id="punchcard-date-rect" class="punchcard-common punchcard-date-rect"></div>
 
-        <script type="text/javascript" src="js/dataloader.js"></script>
-        <script src="js/main-script.js"></script>
+        <script type="text/javascript" src="dataloader.js"></script>
+        <script type="text/javascript" src="main-script.js"></script>
 
   </body>
 </html>
