+++ src/date-circle.ts
@@ -1,8 +1,8 @@
-/// <reference path="../../../typings/globals/crossfilter/index.d.ts" />
-/// <reference path="../../../typings/globals/d3/index.d.ts" />
-/// <reference path="../../../typings/globals/moment/index.d.ts" />
+/// <reference path="../typings/globals/crossfilter/index.d.ts" />
+/// <reference path="../typings/globals/d3/index.d.ts" />
+/// <reference path="../typings/globals/moment/index.d.ts" />
 
-import {IDataRow} from '../idatarow';
+import {IDataRow} from './idatarow';
 import {ColorMap} from './colormap';
 import {DateRect} from './date-rect';
