+++ src/date-rect.ts
@@ -1,9 +1,9 @@
-/// <reference path="../../../typings/globals/crossfilter/index.d.ts" />
-/// <reference path="../../../typings/globals/d3/index.d.ts" />
-/// <reference path="../../../typings/globals/moment/index.d.ts" />
+/// <reference path="../typings/globals/crossfilter/index.d.ts" />
+/// <reference path="../typings/globals/d3/index.d.ts" />
+/// <reference path="../typings/globals/moment/index.d.ts" />
 
 
-import {IDataRow} from '../idatarow';
+import {IDataRow} from './idatarow';
 import {Base} from './base';
 import {ColorMap} from './colormap';
