+++ tsconfig.json
@@ -12,7 +12,7 @@
         "inlineSources": true
     },
     "files": [
-        "./src/ts/punchcards.ts"
+        "./src/punchcards.ts"
     ],
     "exclude": ["sites"]
 }
